requirements.txt
opencv-python==4.*
pillow==10.*
pyyaml==6.*
nbformat==5.*